refactor(cluster_mgrs): move the cluster manager onto the cluster_mgrs package
