+++ docs/architecture/cluster-management-interface-and-implementations.md
@@ -2,7 +2,7 @@
 
 ## Why
 
-1. To improve the architecture of kas-fleet-manager so that when there are changes to OCM, it should only have very small impact on the code base. At the moment, more components are directly depending on OCM than it should be, e.g. [`clusters_mgr.go`](../../internal/kafka/internal/workers/clusters_mgr.go), and this makes it difficult to change the code base if OCM changes.  
+1. To improve the architecture of kas-fleet-manager so that when there are changes to OCM, it should only have very small impact on the code base. At the moment, more components are directly depending on OCM than it should be, e.g. [`clusters_mgr.go`](../../internal/kafka/internal/workers/cluster_mgrs/clusters_mgr.go), and this makes it difficult to change the code base if OCM changes.  
 2. To allow us easily open source the project. With the right interface/contract in place for cluster management, OCM will just become an implementation detail of the interface. New implementations can then be added by us or the upstream community to support different mechanisms to manage OpenShift/Kubernetes clusters.
 
 ## How
@@ -17,4 +17,4 @@ Compared to the current implementation, the main changes are:
 2. Each cluster could have different `Provider` implementation, this information needs to be stored. The `api.Cluster` type is updated to include additional fields to store this information.
 3. A new `ProviderFactory` interface is added which will return the concrete `Provider` implementation based on the `Providertype` value in each `api.Cluster` instance.
 4. Modify the [`ClusterService`](../../internal/kafka/internal/services/clusters.go) class to depend on the new `ProviderFactory`, and remove all dependencies on the OCM client.
-5. Modify the [`ClusterManager`](../../internal/kafka/internal/workers/clusters_mgr.go) class to only depend on the `ClusterService`. It should not have any direct dependencies on the OCM clients.
+5. Modify the [`ClusterManager`](../../internal/kafka/internal/workers/cluster_mgrs/clusters_mgr.go) class to only depend on the `ClusterService`. It should not have any direct dependencies on the OCM clients.
