+++ internal/kafka/internal/workers/cluster_mgrs/clusters_mgr_test.go
@@ -1,4 +1,4 @@
-package workers
+package cluster_mgrs
 
 import (
 	"fmt"
